move diff parsing into own class
