+++ lib/danger_plugin.rb
@@ -1,1 +1,2 @@
 require "todoist/plugin"
+require "todoist/diff_todo_finder"
