+++ spec/todoist_spec.rb
@@ -66,31 +66,6 @@ module Danger
         end
       end
 
-      context "changed files containing removed todos" do
-        before do
-          modified = Git::Diff::DiffFile.new(
-            "base",
-            path:  "some/file.rb",
-            patch: "- TODO: some todo"
-          )
-
-          allow(@dangerfile.git).to receive(:diff_for_file)
-            .with("some/file.rb")
-            .and_return(modified)
-
-          allow(@dangerfile.git).to receive(:modified_files)
-            .and_return([modified_with_todo])
-          allow(@dangerfile.git).to receive(:added_files)
-            .and_return([])
-        end
-
-        it "does not need to warn" do
-          @todoist.warn_for_todos
-
-          expect(warnings).to eq([])
-        end
-      end
-
       context "changed files not containing a todo" do
         before do
           modified = Git::Diff::DiffFile.new(
